fix: add trailingSlash
